Update to latest server packages, rename CarDealership folder
