--- server/CarDealership/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="CarDealership.WebApiApplication" Language="C#" %>
