+++ server/FieldEngineer.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FieldEngineer", "CarDealership\FieldEngineer.csproj", "{1BE26A6D-35F3-4594-AEC5-4DAF9C1BDE20}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FieldEngineer", "FieldEngineer\FieldEngineer.csproj", "{1BE26A6D-35F3-4594-AEC5-4DAF9C1BDE20}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
