+++ server/FieldEngineer/Global.asax.cs
@@ -1,10 +1,9 @@
 ﻿using System.Web.Http;
 using System.Web.Routing;
 using System.Web.Mvc;
-using System.Web.Routing;
 using System.Web.Optimization;
 
-namespace CarDealership
+namespace FieldEngineer
 {
     public class WebApiApplication : System.Web.HttpApplication
     {
