+++ server/FieldEngineer/Views/Admin/Delete.cshtml
