+++ server/FieldEngineer/Views/web.config
@@ -17,7 +17,7 @@
         <add namespace="System.Web.Mvc.Html" />
         <add namespace="System.Web.Routing" />
         <add namespace="System.Web.Optimization" />
-        <add namespace="CarDealership" />
+        <add namespace="FieldEngineer" />
         
       </namespaces>
     </pages>
