+++ server/FieldEngineer/Web.config
@@ -16,16 +16,9 @@
         <!-- Use these settings for local development. After publishing to your
     Mobile App, these settings will be overridden by the values specified
     in the portal. -->
-        <add key="MS_MobileServiceName" value="CarDealership" />
-        <add key="MS_MasterKey" value="Overridden by portal settings" />
-        <add key="MS_ApplicationKey" value="Overridden by portal settings" />
-        <add key="MS_SigningKey" value="Overridden by portal settings" />
-        <add key="EMA_MicroserviceID" value="Overridden by portal settings" />
-        <add key="EMA_RuntimeUrl" value="Overridden by portal settings" />
-        <add key="EMA_Secret" value="Overridden by portal settings" />
+        <add key="MS_MobileServiceName" value="FieldEngineer" />
         <!-- When using this setting, be sure to add matching Notification Hubs connection
     string in the connectionStrings section with the name "MS_NotificationHubConnectionString". -->
-        <add key="MS_NotificationHubName" value="Overridden by portal settings" />
         <add key="webpages:Version" value="3.0.0.0" />
         <add key="webpages:Enabled" value="false" />
         <add key="ClientValidationEnabled" value="true" />
@@ -90,6 +83,10 @@
             <dependentAssembly>
                 <assemblyIdentity name="WebGrease" publicKeyToken="31bf3856ad364e35" culture="neutral" />
                 <bindingRedirect oldVersion="0.0.0.0-1.5.2.14234" newVersion="1.5.2.14234" />
+            </dependentAssembly>
+            <dependentAssembly>
+                <assemblyIdentity name="Microsoft.Azure.Mobile.Server" publicKeyToken="31bf3856ad364e35" culture="neutral" />
+                <bindingRedirect oldVersion="0.0.0.0-1.0.0.0" newVersion="1.0.0.0" />
             </dependentAssembly>
         </assemblyBinding>
     </runtime>
