+++ server/FieldEngineer/packages.config
@@ -8,24 +8,26 @@
   <package id="EntityFramework.SqlServerCompact" version="6.1.3" targetFramework="net452" />
   <package id="jQuery" version="1.10.2" targetFramework="net452" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net452" />
+  <package id="Microsoft.AspNet.Cors" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.Web.Optimization" version="1.1.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net452" />
+  <package id="Microsoft.AspNet.WebApi.Cors" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.OData" version="5.5.1" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Tracing" version="5.2.3" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Authentication" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.CrossDomain" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Entity" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Home" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Notifications" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Quickstart" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.Mobile.Server.Tables" version="0.2.553" targetFramework="net452" />
-  <package id="Microsoft.Azure.NotificationHubs" version="1.0.2" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Authentication" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.CrossDomain" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Entity" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Home" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Notifications" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Quickstart" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.Mobile.Server.Tables" version="1.0.119.0" targetFramework="net452" />
+  <package id="Microsoft.Azure.NotificationHubs" version="1.0.3" targetFramework="net452" />
   <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net452" />
   <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net452" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.3" targetFramework="net452" />
@@ -37,7 +39,7 @@
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
   <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.1.0" targetFramework="net452" />
   <package id="Modernizr" version="2.6.2" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net452" />
   <package id="Owin" version="1.0" targetFramework="net452" />
   <package id="RazorEngine" version="3.7.0" targetFramework="net452" />
   <package id="System.IdentityModel.Tokens.Jwt" version="4.0.2.205111437" targetFramework="net452" />
